Move nested_imageable helper methods to a new common actions module
